[GH-370] 🐛 Fixed the wrongly visible links in the header's navigation bar
The problem was that `height: auto` was set for the `.links` container which
caused the parent container to "overflow", showing the links anyways.
It seems I didn't notice that since the link color was the same as the background.
This is now fixed by setting `overflow: hidden` on the parent `nav` container.