Remove extra multiply by 255
During the debugging session of the coding challenge an extra
multiplication by 255 was added to setting the pixel color. This
is unnecessary and misleading as it simply serves to increase (by a
very large factor) the brightness of the ripples. A better way is to
just drop in higher values for the heights in mouseDragged().