Flocking in JavaFX
Hi :)
I worked on a full Java version of the FlockingSimulation. 
Thinking about it, making this a gradle probject was maybe a bit overhead, but it works relativly smoothly. Since Processing and P5
provide a nice Vector-Implementation but Java not, I also added a small Vector-Datatype, maybe someone will reuse this ^^ It got the relevant methods like sub/add/dist/setMag/normaliez etc.

Was definitly a bit more tricky redoing this in java :)
https://github.com/scorixear/FlockingSimulation