Update 144-black-hole-visualization.md
Something fun I made in P5js. Drawing photon orbits around a spinning black hole. Also the interesting surfaces of the black hole are drawn. 

You can change the parameters of the orbit which are: the spin of the black hole and the radius of the orbit.

url: https://bongo0.github.io/Kerr-orbits/
source: https://github.com/bongo0/Kerr-orbits