add contribution to challenge #132
I tried my best to write this challenge in p5.js. 
Due to the nature of the challenge and the unreadable code of the source the best way to describe what I was doing is captured in this picture https://i.kym-cdn.com/photos/images/original/000/234/765/b7e.jpg. 
Unfortunately I could not locate the bug, and the fluid does not behave properly. 
I spent waaay too much time trying to fix it, so I decided to share it as it is. 

The source https://github.com/bohdandrahan/Fluid-simulation