TSP with Genetic algorithm (#1614)
GA for TSP with a 2-opt move as a mutation and no crossover. As a fitness function I used f(s) = av - len(s), where len - length of the path s, and av is a value more than most of the pathlengths and of the same order. After some constant tweaking I managed to make it work as good as 2-opt and faster than simulated annealing.