ml5-train-your-own contribution: regression added (#1946)
Link to demo: https://editor.p5js.org/raisinten/full/Ho2LJo4Iw

This is a regression-based model. I have also added the mouseDragged functionality.

Special addition:
I have used all the basic notes, i.e, C, D, E, F, G, A and B, which is 7 in number. What else goes with 7 and is colorful too? A rainbow of course! It has 7 colors! So, I mapped the notes with all the 7 colors of the rainbow and topped it with some regression.

Thank you. :)