Add port to P5.js of Coding Challenge #103: Fire Effect
This is a straight port of the current Processing code, making it work
with as few changes as I could manage.

The most obvious and biggest differences are the index calculations for
pixels arrays, and setting 4 instead of 1 entry at a time; both of
which are necessary due to the different structure of these arrays.