Add contribution to CC 103
The main point of this contribution is the explanation (in the source) of the simple optimization it does to speed up the p5.js version. I'm hoping it's explained well enough that people who followed along with the video can figure it out, even without a video showing the process, and thus can learn something in the process.