Contribution:Gift Wrapping Algorithm Challenge#148
Link to Demo:
https://editor.p5js.org/arthurrc/full/OldqKcL5

My take on the challenge where I used vector's cross product from a spinning line whit the lines created from the current point to all the other points. When the z coordinate from the product equals zero, means the lines a parallel. Now, given that i'm rotating at a speed of 0.005 of an angle per frame, there is a precision issue, and I had to make a threshold for the z coordinate absolute value to be close to zero, to consider that line "parallel"... but the result was ok about 80% of the time!