Correct the amount of neighbours cells have in 2D (#2197)
When moving from 3d to 2d the amount of neighbours each cell has changes from 6 to 4. This is needs to be taken into account because lin_solve() calculates the weighted average of the neighbours, and uses the given float c as the total weight. So the provided "c" should be adjusted to correspond with the amount of neighbours.

Without the adjustment lin_solve() reduces, whatever you put into it, by 33% after each iteration.

Co-authored-by: Nico Finkernagel <nico@gruselhaus.com>