Tic Tac Toe Contribution: Made with only Images.
Link: https://editor.p5js.org/ssmkhrj/full/JzpBW6cxO
I have made this Tic Tac Toe game using Minimax algorithm, and by placing a bunch of images on the canvas. Every time the user clicks on a valid board position, a heart image is displayed which is chosen randomly from an array of heart images (to give it a realistic feeling). After that the bot places its move using a cross image which is again a random image from an array of cross images.