Community Contribution - BubbleSort Visualizer
I have created a Bubble Sort Visualizer where the active indices that are being checked for swaps are shown in Green color. And the numbers that have been placed correctly are shown in blue color. Also, added some width to the bars and also these widths can be tweaked by the user, using the slider that is present on the top left corner of the screen. This will unable the user to see bubble sort for varying lengths of the array and observe how the time taken by bubble sort increases quadratically with increase in the length of the array.