Community contribution: Marksweeper by stemvork
Not as pretty as Sayan Goswami's, but I did implement a marking system. I would be honored to be included in the list; it will be my first contribution after years of YouTube enjoyment. Recently worked mostly with pygame, but am also eager to work on a coding challenge in P5.js again.
https://github.com/stemvork/quickle/blob/master/minesweeper/screenshot.png