Small fix in cave generation with marching squares
The `countNeighbors` function counts the number of 1's there are in a 3x3 grid centered at the given (x, y). The condition in question is supposed to check that (i,j) is in a real neighbor and not the node itself ( (x,y) != (i, j)). A small typo slipped in causing it to check that (x, x) != (i, j).