add rel attribute to prevent click-jacking
`target="_blank"` is open to click-jacking attacks, so adding `rel="noopener noreferrer"` is a way to prevent that.

As an aside, both `noopener` and `noreferrer` prevent this vulnerability, though `noopener` won't work with some legacy browsers, and newer browsers just ignore `noreferrer`, so adding both is the current accepted best practice.