Space Invaders contribution : Hanzi Space Invaders (#2704)
Hanzi Space Invaders is space invaders but instead of aliens or asteroids its all Chinese characters! This game was made as a part of several games I made for my local Chinese tuition center to teach young kids Chinese characters. The Coding Train's Space Invaders challenge was a great inspiration for me and was the baseline of this game!

Co-authored-by: Violet <43045568+violetcraze@users.noreply.github.com>