Contribution: pendulum on a cart
As requested by Adam at https://twitter.com/shiffman/status/1363221496018178055
Realistic simulation of a PID controller working on a cart-pendulum system. Limitations: the controller only works on the angle not the position, and adding realistic drag or damping made the differential equations a bit too hard to be much fun anymore :) so, no damping.