Contributing my variant of the prime spiral
Hello,

I would like to add my version of the prime spiral. It was developed using TypeScript and 2D Canvas and can viewed here: https://jackhodgkiss.github.io/prime-spiral/.

It includes the ability to traverse through the spiral with a generation function that uses rows and columns to generate the term as opposed to generating successively from one.